+++ mealie/CHANGELOG.md
@@ -1,4 +1,11 @@
 
+## 0.5.4 (03-12-2021)
+- Update to latest version from hay-kot/mealie
+
+## 0.5.3 (31-10-2021)
+- Update to latest version from hay-kot/mealie
+- Added ssl option
+
 ## 0.5.2 (26-07-2021)
 - Update to latest version from hay-kot/mealie
 - :arrow_up: Initial release